+++ packages/adapter-kysely/src/lucia.d.ts
@@ -1,9 +1,0 @@
-/// <reference types="lucia-auth" />
-declare namespace Lucia {
-	type Auth = any;
-	type UserAttributes = {};
-}
-
-declare namespace App {
-	interface Locals {}
-}